add(I1checklist): add (almost) final changes
This commit represents 99% of the remaining changes that will be done to the
code base before the submission. I added json support for my factory class,
fixed the non-const reference issue in cpplint, and put out a LOT of related
fires.

Remaining issues: make check-style still doesn't run error free. There
is a single issue in a particular file that I am going to ask about in
lab or office hours.